add funcitional options to GET requests to enable setting custom headers and extracting values from the response
this is needed in order to properly make requests for ssz encoded values. we need to have a way to
set headers other than Accept: application/json in the request, and also a way to extract a header from
the response in order to get the version that the beacon state is.